consensus: update dbft, use millisecond-precision time, add CV reason

+++ pkg/consensus/payload_test.go
@@ -224,7 +224,7 @@ func randomMessage(t *testing.T, mt messageType) io.Serializable {
 	switch mt {
 	case changeViewType:
 		return &changeView{
-			timestamp: rand.Uint32(),
+			timestamp: rand.Uint64(),
 		}
 	case prepareRequestType:
 		return randomPrepareRequest(t)
@@ -235,7 +235,7 @@ func randomMessage(t *testing.T, mt messageType) io.Serializable {
 		random.Fill(c.signature[:])
 		return &c
 	case recoveryRequestType:
-		return &recoveryRequest{timestamp: rand.Uint32()}
+		return &recoveryRequest{timestamp: rand.Uint64()}
 	case recoveryMessageType:
 		return randomRecoveryMessage(t)
 	default:
@@ -289,7 +289,7 @@ func randomRecoveryMessage(t *testing.T) *recoveryMessage {
 		},
 		changeViewPayloads: []*changeViewCompact{
 			{
-				Timestamp:          rand.Uint32(),
+				Timestamp:          rand.Uint64(),
 				ValidatorIndex:     3,
 				OriginalViewNumber: 3,
 				InvocationScript:   random.Bytes(4),
